+++ packages/eslint-config-generator/src/config.ts
@@ -1,0 +1,6 @@
+import { Linter } from 'eslint';
+
+export type ESLintConfig = Linter.Config & {
+  extends?: string[];
+  parserOptions: Linter.ParserOptions;
+};
