+++ packages/eslint-config-generator/src/plugin.ts
@@ -1,0 +1,10 @@
+import { ESLintConfig } from './config';
+import { Options } from './options';
+
+export type PluginCallback = ({ config, options }: { config: ESLintConfig; options: Options }) => ESLintConfig;
+
+export function createPlugin(callback: PluginCallback): PluginCallback {
+  return ({ config, options }) => {
+    return callback({ config, options });
+  };
+}
