+++ packages/plugin-vue/__tests__/fixtures/.eslintrc.json
@@ -1,0 +1,3 @@
+{
+  "root": true
+}
