+++ packages/plugin-vue/package.json
@@ -18,7 +18,7 @@
     "rebuild": "npm run build",
     "test": "npm run test:ts && npm run test:unit",
     "test:ts": "tsc --noEmit",
-    "test:unit": "echo 'No unit tests.'",
+    "test:unit": "jest",
     "lint": "eslint . --ext .js,.ts",
     "lint:fix": "npm run lint -- --fix"
   },
@@ -37,7 +37,7 @@
     "eslint-plugin-vue": "^7.1.0"
   },
   "peerDependencies": {
-    "typescript": "^4.0.0",
+    "@kocal/eslint-config-generator": "workspace:*",
     "eslint": "^7.0.0"
   },
   "devDependencies": {
@@ -45,6 +45,8 @@
     "@babel/eslint-parser": "^7.12.1",
     "@babel/preset-env": "^7.12.1",
     "@babel/preset-typescript": "^7.12.1",
+    "@kocal/eslint-config-generator": "workspace:*",
+    "@kocal/eslint-config-generator-plugin-typescript": "workspace:*",
     "@rollup/plugin-babel": "^5.2.1",
     "@rollup/plugin-node-resolve": "^9.0.0",
     "@types/eslint": "^7.2.4",
